pyproject.toml
[tool.poetry]
name = "finance-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
psycopg2-binary = "^2.9.10"
django-tailwind = {extras = ["reload"], version = "^3.8.0"}
django-allauth = "^65.3.1"
django-debug-toolbar = "^4.4.6"
django-extensions = "^3.2.3"
django-filter = "^24.3"
django-widget-tweaks = "^1.5.0"
requests = "^2.32.3"
faker = "^33.1.0"
heroicons = {extras = ["django"], version = "^2.9.0"}
django-htmx = "^1.21.0"
pytest = "^8.3.4"
pytest-django = "^4.9.0"
factory-boy = "^3.3.1"
django-template-partials = "^24.4"
plotly = "^5.24.1"
pandas = "^2.2.3"
django-import-export = "^4.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
